feat: 좋아요 관련 api 생성 및 기존 목록 불러올 시 좋아요 여부, 개수도 불러오도록 수정 #3

+++ src/main/java/net/ohmwomuc/domain/likes/service/LikesServiceImpl.java
@@ -6,6 +6,8 @@ import net.ohmwomuc.domain.likes.repository.LikesRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Objects;
+
 @Service
 @RequiredArgsConstructor
 public class LikesServiceImpl implements LikesService {
@@ -19,18 +21,31 @@ public class LikesServiceImpl implements LikesService {
 
     @Override
     @Transactional
-    public Like.DomainResponse addLikesCountByMuamucId(Like.DomainRequest likeRequest) {
-        likesRepository.addLikesCount(likeRequest);
-        Like.DomainRequest output = Like.DomainRequest.builder()
-                .muamuc_id(likeRequest.getMuamuc_id())
-                .user_id(likeRequest.getUser_id())
-                .build();
-        return likesRepository.getLikesByuserIdAndMuamucId(output);
+    public Boolean reverseLike(Like.DomainRequest likeRequest) {
+        if (Objects.isNull(likesRepository.getLikesByuserIdAndMuamucId(likeRequest))) {
+            likesRepository.addLikesCount(likeRequest);
+            return true;
+        } else {
+            likesRepository.deleteLikesCount(likeRequest);
+            return false;
+        }
+
     }
 
-    @Override
-    @Transactional
-    public void deleteLike(Like.DomainRequest likeRequest) {
-        likesRepository.deleteLikesCount(likeRequest);
-    }
+//    @Override
+//    @Transactional
+//    public Like.DomainResponse addLikesCountByMuamucId(Like.DomainRequest likeRequest) {
+//        likesRepository.addLikesCount(likeRequest);
+//        Like.DomainRequest output = Like.DomainRequest.builder()
+//                .muamuc_id(likeRequest.getMuamuc_id())
+//                .user_id(likeRequest.getUser_id())
+//                .build();
+//        return likesRepository.getLikesByuserIdAndMuamucId(output);
+//    }
+//
+//    @Override
+//    @Transactional
+//    public void deleteLike(Like.DomainRequest likeRequest) {
+//        likesRepository.deleteLikesCount(likeRequest);
+//    }
 }
